Split into two crates, bump to v0.2.0
