+++ Cargo.toml
@@ -1,13 +1,2 @@
-[package]
-name = "acpi"
-version = "0.1.0"
-authors = ["Isaac Woods <isaacwoods.home@gmail.com>"]
-repository = "https://github.com/rust-osdev/acpi"
-description = "Library for parsing ACPI tables and AML"
-readme = "README.md"
-license = "MIT/Apache-2.0"
-edition = "2018"
-
-[dependencies]
-log = "0.*"
-bit_field = "0.9.0"
+[workspace]
+members = ["acpi", "aml_parser"]
