+++ acpi/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "acpi"
+version = "0.2.0"
+authors = ["Isaac Woods"]
+repository = "https://github.com/rust-osdev/acpi"
+description = "Library for parsing ACPI tables"
+categories = ["hardware-support", "no-std"]
+readme = "../README.md"
+license = "MIT/Apache-2.0"
+edition = "2018"
+
+[dependencies]
+log = "0.4"
+bit_field = "0.9"
