+++ aml_parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "aml_parser"
+version = "0.2.0"
+authors = ["Isaac Woods"]
+repository = "https://github.com/rust-osdev/acpi"
+description = "Library for parsing AML"
+categories = ["hardware-support", "no-std"]
+readme = "../README.md"
+license = "MIT/Apache-2.0"
+edition = "2018"
+
+[dependencies]
